requirements.txt
wheel>=0.37.1
cython>=0.29.26
astropy>=5.0
astroplan>=0.8
astroquery>=0.4.5
backports-abc>=0.5
bokeh==2.4.2
certifi>=2021.10.8
Django>=4.0.1
djangorestframework>=3.13.1
djangorestframework-datatables>=0.7.0
django-filter>=21.1
ephem>=4.1.3
futures==3.1.1
h5py>=3.6.0
Jinja2>=3.0.3
MarkupSafe>=2.0.1
numpy>=1.22.0
pyparsing>=3.0.6
python-dateutil>=2.8.2
pytz>=2021.3
PyYAML>=6.0
requests>=2.27.1
scipy>=1.7.3
singledispatch>=3.7.0
six>=1.16.0
tornado>=6.1
pandas>=1.3.5
gunicorn>=20.1.0
psycopg2-binary>=2.9.3
django-environ>=0.8.1
specutils>=1.5.0
